op-acceptance_tests: consolidate hardforks_ext and sync_tester_hfs_ext
